+++ packages/world-vercel/src/storage.ts
@@ -9,6 +9,7 @@ import { instrumentObject } from './instrumentObject.js';
 import {
   experimentalSetAttributes,
   getWorkflowRun,
+  getWorkflowRuns,
   listWorkflowRuns,
 } from './runs.js';
 import { getStep, listWorkflowRunSteps } from './steps.js';
@@ -20,6 +21,10 @@ export function createStorage(config?: APIConfig): Storage {
     runs: {
       get: ((id: string, params?: any) =>
         getWorkflowRun(id, params, config)) as Storage['runs']['get'],
+      getMany: ((ids: readonly string[], params?: any) =>
+        getWorkflowRuns(ids, params, config)) as NonNullable<
+        Storage['runs']['getMany']
+      >,
       list: ((params?: any) =>
         listWorkflowRuns(params, config)) as Storage['runs']['list'],
       experimentalSetAttributes: (runId, changes, options) =>